Reoganize main function #188 - displaced ng related code in specific file
